Fixing delete area
