+++ play/src/front/Phaser/Game/MapEditor/Tools/TrashEditorTool.ts
@@ -9,21 +9,16 @@ import { EntityRelatedEditorTool } from "./EntityRelatedEditorTool";
 
 export class TrashEditorTool extends EntityRelatedEditorTool {
     private areaPreviews: AreaPreview[];
-
     protected ctrlKey?: Phaser.Input.Keyboard.Key;
-
     private active = false;
-
-    // private mapEditorSelectedAreaPreviewStoreSubscribe!: Unsubscriber;
 
     constructor(mapEditorModeManager: MapEditorModeManager) {
         super(mapEditorModeManager);
 
         this.active = false;
-
         this.ctrlKey = this.scene.input.keyboard?.addKey("CTRL");
-
         this.areaPreviews = this.createAreaPreviews();
+        this.bindEventHandlers();
     }
 
     private createAreaPreviews(): AreaPreview[] {
@@ -40,7 +35,7 @@ export class TrashEditorTool extends EntityRelatedEditorTool {
     }
 
     private createAreaPreview(areaConfig: AreaData): AreaPreview {
-        const areaPreview = new AreaPreview(this.scene, structuredClone(areaConfig), this.shiftKey, this.ctrlKey);
+        const areaPreview = new AreaPreview(this.scene, structuredClone(areaConfig), undefined, this.ctrlKey);
         this.bindAreaPreviewEventHandlers(areaPreview);
         this.areaPreviews.push(areaPreview);
         return areaPreview;
@@ -166,16 +161,12 @@ export class TrashEditorTool extends EntityRelatedEditorTool {
     }
 
     protected bindEventHandlers(): void {
-        super.bindEventHandlers();
-
         this.scene.input.on(Phaser.Input.Events.POINTER_UP, this.pointerUpEventHandler);
         this.scene.input.on(Phaser.Input.Events.POINTER_OVER, this.pointerHoverEventHandler);
         this.scene.input.on(Phaser.Input.Events.POINTER_OUT, this.pointerOutEventHandler);
     }
 
     protected unbindEventHandlers(): void {
-        super.unbindEventHandlers();
-
         this.scene.input.off(Phaser.Input.Events.POINTER_UP, this.pointerUpEventHandler);
         this.scene.input.off(Phaser.Input.Events.POINTER_OVER, this.pointerHoverEventHandler);
         this.scene.input.off(Phaser.Input.Events.POINTER_OUT, this.pointerOutEventHandler);
@@ -195,10 +186,6 @@ export class TrashEditorTool extends EntityRelatedEditorTool {
         console.info("handleAreaPreviewCreation => No create area preview in trash mode");
     }
 
-    public async handleIncomingCommandMessage(editMapCommandMessage: EditMapCommandMessage): Promise<void> {
-        // Nothing to do here
-    }
-
     public activate() {
         super.activate();
         this.active = true;
@@ -214,8 +201,7 @@ export class TrashEditorTool extends EntityRelatedEditorTool {
         this.scene.markDirty();
     }
 
-    protected bindEntitiesManagerEventHandlers(): void {
-        // NOTE: We don't need to bind anything here, because all the events emitted by EntitiesManager are handled by
-        // the parent of this class (EntityEditorTool) who is already called in the MapEditorModeManager constructor.
+    handleIncomingCommandMessage(editMapCommandMessage: EditMapCommandMessage): Promise<void> {
+        return Promise.resolve(undefined);
     }
 }
